GitHub Health Files: Contributing
As part of the community health files [1], create a contribution
guidelines [2] to communicate to project/repository contributors [3].
The feature is configured by a `contributing.md` file [4] that must be
placed either in the root, `/docs` or `.github` directory of an
repository.

For Nord this file is provided globally for all projects by this
organization `.github` repository that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://en.wikipedia.org/wiki/Contributing_guidelines
[3]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
[4]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file

GH-6