GitHub Health Files: Security Policy
As part of the community health files [1], a security policy [2] to give
instructions for how to report a security vulnerability the project [3]
has been created.
The feature is configured by a `security.md` file that must be placed
either in the root, `/docs` or `.github` directory of an repository.

For Nord this file is provided globally for all projects by this
organization `.github` repository that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://en.wikipedia.org/wiki/Security_policy
[3]: https://docs.github.com/en/code-security/getting-started/adding-a-security-policy-to-your-repository

GH-8