GitHub Health Files: Sponsors (Funding)
As part of the community health files [1], the GitHub sponsors [2]
configuration file file added to display a sponsor button in an
repository [3].
The feature is configured by the `funding.yml` file that must be placed
either in the root, `/docs` or `.github` directory of an repository.

For Nord this file is provided globally by this organization's `.github`
repository for all repositories that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://github.com/sponsors
[3]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/displaying-a-sponsor-button-in-your-repository

GH-10