GitHub Health Files: Issue Template Chooser
As part of the community health files [1], the issue template chooser
configuration file [2] to customize the issue template chooser that is
shown when creating a new issue in a repository.
The feature is configured by the `issue_template/config.yml` file that
must be placed either in the root, `/docs` or `.github` directory of a
repository.

For Nord this file is provided globally by this organization's `.github`
repository for all repositories that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#configuring-the-template-chooser

GH-12