GitHub Health Files: Support Resources
As part of the community health files [1], the support resources [2]
file to provide instructions for how to ask for help and general Nord
community & support resources.
The feature is configured by the `support.md` file [3] that must be
placed either in the root, `/docs` or `.github` directory of a
repository.

For Nord this file is provided globally by this organization's `.github`
repository for all repositories that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-support-resources-to-your-project
[3]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-support-resources-to-your-project#adding-support-resources-to-your-project

GH-14